fix(BA-7742): keep the scope item type out of the rbac import cycle
`UUIDScopeGQL` sat under `rbac`, so importing it ran that package's `__init__`,
which reaches back through the entity node into resource groups and deployments.
Any module in that reach could no longer be imported on its own. The type is a
shared GraphQL primitive rather than an rbac one, so it moves to `gql/base.py`;
the resource group's scope input moves out of `types.py` for the same reason.
