+++ src/ai/backend/manager/api/gql/entity_share/types.py
@@ -38,6 +38,7 @@ from ai.backend.common.dto.manager.v2.entity_share.types import (
     EntityShareStatusDTO,
 )
 from ai.backend.common.meta.meta import NEXT_RELEASE_VERSION
+from ai.backend.manager.api.gql.base import UUIDScopeGQL
 from ai.backend.manager.api.gql.decorators import (
     BackendAIGQLMeta,
     PydanticInputMixin,
@@ -49,7 +50,7 @@ from ai.backend.manager.api.gql.decorators import (
     gql_pydantic_type,
 )
 from ai.backend.manager.api.gql.pydantic_compat import PydanticNodeMixin, PydanticOutputMixin
-from ai.backend.manager.api.gql.rbac.types.scope import PermissionBitGQL, UUIDScopeGQL
+from ai.backend.manager.api.gql.rbac.types.scope import PermissionBitGQL
 
 EntityShareStatusGQL: type[EntityShareStatusDTO] = gql_enum(
     BackendAIGQLMeta(added_version=NEXT_RELEASE_VERSION, description="Where a share stands."),
